Add a Vary header to content negotation endpoints
When the browser requests `/signals/` etc and gets json back this wipes
out the cache of the html. Then if you press back the browser fetches
the content from the cache which is now a json file.

If the content varys on the `Accept` header then the browser can store
the two different versions in the cache.